requirements.txt
altgraph==0.17
appdirs==1.4.4
astroid==2.4.2
attrs==20.1.0
black==19.10b0
click==7.1.2
isort==4.3.21
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pathspec==0.8.0
pkg-resources==0.0.0
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.7
pylint==2.5.3
PyYAML==5.3.1
regex==2020.7.14
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
